fix: fix calculating mark algorithm

+++ StudyHub/Entities/Assignment.cs
@@ -5,7 +5,7 @@ namespace StudyHub.Entities;
 public class Assignment : EntityBase
 {
     public string Title { get; set; } = string.Empty;
-    public int MaxMark { get; set; }
+    public double MaxMark { get; set; }
     public DateTime OpeningDate { get; set; }
     public DateTime ClosingDate { get; set; }
     public TimeSpan Duration { get; set; }
